hotfix part Float

+++ threshold_public_goods_end/pages.py
@@ -97,8 +97,8 @@ class Summary(Page):
 				TokensEarnedP2 = str(self.player.participant.vars['rec'])
 				self.player.round_chosen_for_payment_P2=self.player.participant.vars['keptR']
 				self.player.Part2Earnings=ME2
-				ME2 = int(TokensKeptP2)+int(TokensEarnedP2)
-		self.player.payoff=int(ME1)+int(ME2)+int(self.session.config['participation_payment'])
+				ME2 = float(TokensKeptP2)+float(TokensEarnedP2)
+		self.player.payoff=float(ME1)+float(ME2)+float(self.session.config['participation_payment'])
 		return dict(
 			d,
 			all_vars = self.participant.vars,
